# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: WorkerConnect API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workers/register:
    post:
      operationId: registerWorker
      tags: [workers]
      summary: Register a new worker
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkerRegistration"
      responses:
        "201":
          description: Worker registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Worker"
        "400":
          description: Validation error
  /workers/me:
    get:
      operationId: getWorkerProfile
      tags: [workers]
      summary: Get current worker profile
      responses:
        "200":
          description: Worker profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Worker"
    patch:
      operationId: updateWorkerProfile
      tags: [workers]
      summary: Update current worker profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkerUpdate"
      responses:
        "200":
          description: Updated worker
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Worker"
  /workers/me/availability:
    patch:
      operationId: toggleAvailability
      tags: [workers]
      summary: Toggle worker online/offline status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AvailabilityInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Worker"
  /workers/me/fcm-token:
    post:
      operationId: updateFcmToken
      tags: [workers]
      summary: Register FCM push notification token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FcmTokenInput"
      responses:
        "200":
          description: Token saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /payouts:
    get:
      operationId: listPayouts
      tags: [payouts]
      summary: List all payouts (admin)
      parameters:
        - name: workerId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Payout list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payout"
    post:
      operationId: triggerPayout
      tags: [payouts]
      summary: Trigger a payout to a worker (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PayoutInput"
      responses:
        "201":
          description: Payout triggered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payout"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # ── Auth ──────────────────────────────────────────────────────────────────
...
    Worker:
      type: object
      required: [id, name, phone, city, skill, subscriptionStatus, isOnline, isSuspended, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        city:
          type: string
        skill:
          type: string
        upiId:
          type: ["string", "null"]
        profileImage:
          type: ["string", "null"]
        isOnline:
          type: boolean
        isSuspended:
          type: boolean
        subscriptionStatus:
          type: string
        fcmToken:
          type: ["string", "null"]
        createdAt:
          type: string
    WorkerRegistration:
      type: object
      required: [name, phone, city, skill]
      properties:
        name:
          type: string
        phone:
          type: string
        city:
          type: string
        skill:
          type: string
        upiId:
          type: string
        profileImage:
          type: string
    WorkerUpdate:
      type: object
      properties:
        name:
          type: string
        city:
          type: string
        upiId:
          type: string
        profileImage:
          type: string
    AvailabilityInput:
      type: object
      required: [isOnline]
      properties:
        isOnline:
          type: boolean
    FcmTokenInput:
      type: object
      required: [fcmToken]
      properties:
        fcmToken:
          type: string
...
    Job:
      type: object
      required: [id, title, description, skillRequired, location, payoutAmount, status, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        skillRequired:
          type: string
        location:
          type: string
        payoutAmount:
          type: number
        status:
          type: string
        completionNotes:
          type: ["string", "null"]
        assignedWorkerId:
          type: ["integer", "null"]
        createdAt:
          type: string
...
    Payout:
      type: object
      required: [id, workerId, jobId, amount, status, createdAt]
      properties:
        id:
          type: integer
        workerId:
          type: integer
        jobId:
          type: integer
        amount:
          type: number
        status:
          type: string
        transactionReference:
          type: ["string", "null"]
        paidAt:
          type: ["string", "null"]
        createdAt:
          type: string
        worker:
          $ref: "#/components/schemas/Worker"
        job:
          $ref: "#/components/schemas/Job"
    PayoutInput:
      type: object
      required: [workerId, jobId, amount]
      properties:
        workerId:
          type: integer
        jobId:
          type: integer
        amount:
          type: number

    # ── Analytics ─────────────────────────────────────────────────────────────
...