# lib/api-zod/src/generated/types/payoutInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WorkerConnect API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PayoutInput {
  workerId: number;
  jobId: number;
  amount: number;
}