# lib/api-zod/src/generated/types/workerRegistration.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WorkerConnect API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WorkerRegistration {
  name: string;
  phone: string;
  city: string;
  skill: string;
  upiId?: string;
  profileImage?: string;
}