# lib/api-zod/src/generated/types/workerUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WorkerConnect API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WorkerUpdate {
  name?: string;
  city?: string;
  upiId?: string;
  profileImage?: string;
}